refactor: DRY cleanup — centralize utils, extract components, remove clutter
- Centralize date formatting into lib/date-utils.ts (formatRelativeTime, formatTime, formatFullDateTime, formatDate)
- Remove 5 duplicate formatDate/formatTimestamp/formatMessageTime declarations across pages
- Extract shared inbox components (ChatHeader, ChatInput, ConversationList, MessageList)
- Extract shared notification components (NotificationItem, NotificationsEmptyState, notification-config)
- Move demo data to lib/demo-data.ts (job-seeker + employer)
- Extract Conversation & Message types to types/inbox.ts
- Refactor employer inbox page from 822 to 351 lines using shared components
- Rename employer/messages route to employer/inbox for consistency
- Remove unused ResponsiveTableExample component
- Update DashboardNavbar links to match new routes
