+++ app/dashboard/employer/drafts/page.tsx
@@ -9,13 +9,8 @@ import {
   HiOutlineCalendar,
   HiOutlineArrowRight,
 } from "react-icons/hi2";
+import { formatDate } from "@/lib/date-utils";
 
-function formatDate(value?: string) {
-  if (!value) return "-";
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return "-";
-  return date.toLocaleDateString();
-}
 
 export default function ManageDraftsPage() {
   const { data: jobs = [], isLoading } = useSWR(
