+++ app/dashboard/employer/notifications/page.tsx
@@ -1,57 +1,11 @@
 "use client";
 
 import React from "react";
-import {
-  HiCheckCircle,
-  HiOutlineUserPlus,
-  HiOutlineChatBubbleLeftRight,
-  HiOutlineCheck,
-  HiOutlineXMark,
-} from "react-icons/hi2";
+import { HiOutlineCheck } from "react-icons/hi2";
 import { useNotifications } from "@/hooks/useNotifications";
 import { RoleGuard } from "@/components/auth/RoleGuard";
-
-// ─── Types ───────────────────────────────────────────────────────────────────
-
-type NotificationType =
-  | "application_received"
-  | "application_reviewed"
-  | "system";
-
-// ─── Icon + color config ─────────────────────────────────────────────────────
-
-const NOTIFICATION_CONFIG: Record<
-  NotificationType,
-  {
-    icon: typeof HiCheckCircle;
-    iconColor: string;
-    borderColor: string;
-  }
-> = {
-  application_reviewed: {
-    icon: HiCheckCircle,
-    iconColor: "text-green-500",
-    borderColor: "border-l-green-500",
-  },
-  application_received: {
-    icon: HiOutlineUserPlus,
-    iconColor: "text-primary",
-    borderColor: "border-l-primary",
-  },
-  system: {
-    icon: HiOutlineChatBubbleLeftRight,
-    iconColor: "text-gray-500",
-    borderColor: "border-l-gray-400",
-  },
-};
-
-function formatTimestamp(value: string): string {
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return "Recently";
-  return date.toLocaleString();
-}
-
-// ─── Page ────────────────────────────────────────────────────────────────────
+import { NotificationItem } from "@/components/dashboard/notifications/NotificationItem";
+import { NotificationsEmptyState } from "@/components/dashboard/notifications/NotificationsEmptyState";
 
 export default function NotificationsPage() {
   const {
@@ -93,63 +47,17 @@ export default function NotificationsPage() {
         )}
 
         <div className="space-y-2">
-          {employerNotifications.map((notif) => {
-            const config =
-              NOTIFICATION_CONFIG[notif.type as NotificationType] ||
-              NOTIFICATION_CONFIG.system;
-            const Icon = config.icon;
-
-            return (
-              <div
-                key={notif._id}
-                className={`bg-white rounded-lg border border-gray-200 border-l-4 ${config.borderColor} px-5 py-4 flex items-start gap-3 ${
-                  !notif.is_read ? "bg-yellow-50/40" : ""
-                }`}
-                onClick={() => {
-                  if (!notif.is_read) {
-                    void markRead(notif._id);
-                  }
-                }}
-              >
-                {/* Icon */}
-                <Icon
-                  className={`w-5 h-5 mt-0.5 shrink-0 ${config.iconColor}`}
-                />
-
-                {/* Content */}
-                <div className="flex-1 min-w-0">
-                  <h3 className="text-sm font-semibold text-gray-900">
-                    {notif.title}
-                  </h3>
-                  <p className="text-sm text-gray-500 mt-0.5">
-                    {notif.description}
-                  </p>
-                </div>
-
-                {/* Timestamp */}
-                <span className="text-xs text-gray-400 whitespace-nowrap shrink-0 mt-0.5">
-                  {formatTimestamp(notif.createdAt)}
-                </span>
-
-                <button
-                  type="button"
-                  aria-label="Dismiss notification"
-                  onClick={(event) => {
-                    event.stopPropagation();
-                    void removeNotification(notif._id);
-                  }}
-                  className="ml-1 text-gray-400 hover:text-gray-600"
-                >
-                  <HiOutlineXMark className="w-4 h-4" />
-                </button>
-              </div>
-            );
-          })}
+          {employerNotifications.map((notif) => (
+            <NotificationItem
+              key={notif._id}
+              notification={notif}
+              onMarkAsRead={(id) => void markRead(id)}
+              onDelete={(id) => void removeNotification(id)}
+            />
+          ))}
 
           {!isLoading && employerNotifications.length === 0 && (
-            <div className="bg-white rounded-lg border border-gray-200 p-8 text-center text-sm text-gray-500">
-              No notifications yet.
-            </div>
+            <NotificationsEmptyState />
           )}
         </div>
       </div>
