+++ app/dashboard/job-seeker/applied-jobs/loading.tsx
