+++ app/dashboard/job-seeker/jobs/loading.tsx
