+++ app/dashboard/job-seeker/profile/page.tsx
@@ -720,7 +720,7 @@ export default function ProfilePage() {
                 <div className="flex items-center justify-between mb-4">
                   <div className="flex items-center gap-2">
                     <HiOutlineBriefcase className="w-5 h-5 text-red-500" />
-                    <h3 className="text-lg font-bold text-gray-900">
+                    <h3 className="text-lg md:text-xl font-bold text-gray-900">
                       Professional Information
                     </h3>
                   </div>
