+++ components/pages/dashboard/EmployerDashboard.tsx
@@ -17,13 +17,8 @@ import {
   HiOutlineChevronLeft,
   HiOutlineChevronRight,
 } from "react-icons/hi2";
+import { formatDate } from "@/lib/date-utils";
 
-function formatDate(value?: string) {
-  if (!value) return "—";
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return "—";
-  return date.toLocaleDateString();
-}
 
 // ─── Component ───────────────────────────────────────────────────────────────
