+++ components/ui/EducationCard.tsx
@@ -52,7 +52,9 @@ export const EducationCard: React.FC<EducationCardProps> = ({
       <div className="flex items-center justify-between mb-6">
         <div className="flex items-center gap-2">
           <HiOutlineAcademicCap className="w-5 h-5 md:w-6 md:h-6 text-red-500" />
-          <h2 className="text-xl md:text-3xl font-bold text-gray-900">Education</h2>
+          <h2 className="text-lg md:text-xl font-bold text-gray-900">
+            Education
+          </h2>
         </div>
         <button
           onClick={onAdd}
