+++ components/ui/WorkExperienceCard.tsx
@@ -65,7 +65,9 @@ export const WorkExperienceCard: React.FC<WorkExperienceCardProps> = ({
       <div className="flex items-center justify-between mb-6">
         <div className="inline-flex items-center gap-2">
           <HiOutlineBriefcase className="w-5 h-5 md:w-6 md:h-6 text-red-500" />
-          <h2 className="text-xl md:text-3xl font-bold text-gray-900">Work Experience</h2>
+          <h2 className="text-lg md:text-xl font-bold text-gray-900">
+            Work Experience
+          </h2>
         </div>
         <button
           onClick={onAdd}
