+++ hooks/useSavedSearches.ts
@@ -74,7 +74,7 @@ export function buildSearchUrl(filters: SavedSearchFilters): string {
   if (filters.experienceLevel)
     params.set("experience", filters.experienceLevel);
   const qs = params.toString();
-  return qs ? `/dashboard/jobs?${qs}` : "/dashboard/jobs";
+  return qs ? `/dashboard/job-seeker/jobs?${qs}` : "/dashboard/job-seeker/jobs";
 }
 
 // ─── Hook ───────────────────────────────────────────────────────────────────────
